Move Encodable/Decodable string impls to impls module
Now we have the private `impls` module move the various string
implementations of `Encodable` and `Decodable` into it.
